!!WIP!! sync prefix on startup
